refactor(web/routes): extract dashboard/auth route module, finalize server.ts aggregator
server.ts reduced from 718 → 200 lines. New modules:
- src/web/routes/admin.ts — all /admin/* handlers (adminGuard applied internally)
- src/web/routes/dashboard.ts — /health, /manifest.json, /static/styles.css, auth, GET /

getConversationUrl re-exported from server.ts for backward compatibility.
